db/requirements.txt
psycopg2-binary
qdrant-client
sentence-transformers
requests
tqdm
